feat(RunFrame): add support for custom main component path && updated eval (#542)
* feat(RunFrame): add support for custom main component path

This commit introduces a new `mainComponentPath` prop to the `RunFrame` component, allowing users to specify a custom path for the main component to be rendered. If not provided, the default component from the entry point will be used. A new example fixture has also been added to demonstrate this functionality.

* chore: update @tscircuit/eval dependency to v0.0.181

This commit updates the @tscircuit/eval package from version 0.0.179 to 0.0.181 to ensure the latest features and bug fixes are included in the project.